Code Cleanup & Bugfix

+++ CowGraph/CowGraphModel/src/zbeans/cowgraph/model/GraphElement.java
@@ -44,15 +44,16 @@ public interface GraphElement extends PropertyChangeObservable, Cloneable {
 
     public GraphElementType getType();
 
-    public double getX();
+    public int getX();
 
-    public double getY();
+    public int getY();
 
     public Color getColor();
 
-    public void setX(double x);
+    public void setX(int x);
 
-    public void setY(double y);
-    
+    public void setY(int y);
+
     public GraphElement clone();
+
 }
